Use same component for all user tables
